test_project/requirements.txt
Django==1.11.9
pytz==2017.2
django-model-utils==3.0.0
django-positions==0.5.4
djangorestframework==3.7.0
Pillow==4.1.1
psycopg2==2.7.1
celery==3.1.25
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
django-jet==1.0.6
django-celery==3.2.1
django-polymorphic
django-phonenumber-field
pytest==3.2.1
pytest-django==3.1.2
pytest-cov==2.4.0
mock==2.0.0
django-mock-queries==0.0.15
freezegun==0.3.8
pytest-faker==2.0.0
django-extensions==1.9.0
easy-thumbnails==2.4.2
django-rest-swagger==2.1.2
django-redis==4.8.0
django-debug-toolbar
django-filter==1.1.0
sqlparse
unidecode
fmoo-audiotools==3.1b1
